+++ hebiros/msg/CommandMsg.msg
@@ -1,0 +1,6 @@
+string[] name
+float64[] position
+float64[] velocity
+float64[] effort
+hebiros/SettingsMsg settings
+
