+++ hebiros/msg/PidGainsMsg.msg
@@ -1,0 +1,16 @@
+string[] name
+float64[] kp
+float64[] ki
+float64[] kd
+float64[] feed_forward
+float64[] dead_zone
+float64[] i_clamp
+float64[] punch
+float64[] min_target
+float64[] max_target
+float64[] target_lowpass
+float64[] min_output
+float64[] max_output
+float64[] output_lowpass
+bool[] d_on_error
+
