+++ hebiros/srv/SendCommandWithAcknowledgementSrv.srv
@@ -1,0 +1,2 @@
+CommandMsg command
+---
