+++ hebiros_advanced_examples/launch/other/example_set_gains.launch
@@ -1,0 +1,6 @@
+<launch>
+  <node name="hebiros_node" pkg="hebiros" type="hebiros_node" respawn="false" output="screen"/>
+
+  <node name="example_set_gains_node" pkg="hebiros_advanced_examples" type="example_set_gains_node" respawn="false" output="screen"/>
+
+</launch>
