+++ hebiros_advanced_examples/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>hebiros_advanced_examples</name>
-  <version>1.0.0</version>
+  <version>1.1.0</version>
   <description>Advanced examples for hebiros with additional dependencies</description>
 
   <maintainer email="x@hebirobotics.com">Xavier Artache</maintainer>
